Stopgap -- enable typescript parsing
Just enable the typescript plugin every time when running
extract-messages.

Not sure if there are problems when using it with normal JSX files.
The parser plugins (or all parser options) should probably be
configurable anyway (but the configuration doesn't work for
extract-messages at the moment)

Also passing sourceFileName to babylon.parse, though not sure if it has
any effect.